chore: rename component folders to plural
